Update iOS Deployment Target and Add Camera Usage Description
Update the iOS deployment target to version 13.0 across all build configurations
in the Xcode project file. This change ensures compatibility with the latest iOS
features and improvements. Additionally, add a camera usage description in the
Info.plist file to request permission for accessing the camera, which is necessary
for scanning cards within the application. The description informs the user that
the camera is used to add cards automatically by scanning them.